requirements.txt
aiogram==3.4.1
telethon==1.34.0
aiosqlite==0.19.0
python-dotenv==1.0.0
cryptography==42.0.2
pytgcalls==3.0.0.dev24
